requirements.txt
# Core OCR and Image Processing
pytesseract==0.3.10
easyocr==1.7.0
Pillow==10.1.0
pdf2image==1.16.3

# Vector Database and ML
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# Web Framework and API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# LLM Integration (Optional)
openai==1.3.7

# Utilities
requests==2.31.0
python-dotenv==1.0.0
pathlib